rearrange code, move service deployment to client
