+++ src/main/java/pl/net/gazda/common/service/IgniteService.java
@@ -1,0 +1,7 @@
+package pl.net.gazda.common.service;
+
+import org.apache.ignite.services.Service;
+
+public interface IgniteService extends Service {
+    void performSomeAction();
+}
